Community action shifted to community page and fixed the loaders.
